Update test_creation.py
Added lines for project specific folder which contains final documentation, project plan, and technical debt documents. I think in practice having these all together will help keep teams on the same page in terms of scope and goal of project, keep track of proposed improvements that got away, and enable smooth transitions of work. See below for apt descriptions for these files. 


o	Technical debt (document) – Document detailing the accrued technical debt (i.e. the ideas for improvements not implemented but thought of). This document should be modified during the lifecycle of the efforts and revisited routinely to ensure significant model improvement ideas are implemented in a manner that is congruent to current work, and that these significant ideas are not forgotten about.

 Final Documentation (word document)
		- Document containing:		
1- Executive Summary: summary of projects goals, approach, and results.
			2- Objective: specific business question/need being addressed.
			3- Background: any information on what has preceded this project
			4- Approach: hypothesis and analytical strategy
5- Data: describe where the data came from, what it contained, what was used, and what was done to it.
6- Results: what metric was used, why, and how the model performed.
7- Recommendations: how can the model(s) be improved in the future and what are the implications/uses of the model in the context of businesses needs/functionalities, and uses cases tailored to fit those needs/functionalities.
8- Appendix: exhaustive details of feature engineering and models that were tried and failed.

- Project plan (word document)
	- Document containing:
1- Project description: who started it, what is the business question being asked, and what are the deliverables.
2- Timeline and milestones: where is the project now, when/where has it been, and when/where is it going.
3- Approach: what are you trying and why.
4- Failures: what was tried and failed or not pursued.